Renamed file names to match the class names
