+++ FastCache/FastCache.csproj
@@ -264,10 +264,10 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Configuration.resx</DependentUpon>
     </Compile>
-    <Compile Include="Events.cs" />
+    <Compile Include="FastCacheEvents.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="FastCacheCore.cs" />
-    <Compile Include="Watcher.cs" />
+    <Compile Include="StreamWatcher.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="packages.config" />
