+++ FastCache/StreamWatcher.cs
